@@ -4,13 +4,14 @@ import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Environment;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.app.exptracker.R;
 import com.app.exptracker.utility.Utils;
@@ -19,7 +20,7 @@ import java.io.File;
 import java.util.Collections;
 import java.util.List;
 
-import static android.support.v4.content.FileProvider.getUriForFile;
+import static androidx.core.content.FileProvider.getUriForFile;
 
 public class FileBackupAdapter extends RecyclerView.Adapter<FileBackupAdapter.MyViewHolder> {
 
