@@ -1,7 +1,8 @@
 package com.app.exptracker.utility;
 
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.helper.ItemTouchHelper;
+
+import androidx.recyclerview.widget.ItemTouchHelper;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.app.exptracker.adapter.SortCategoryAdapter;
 
