@@ -1,7 +1,9 @@
 package com.app.exptracker.utility;
 
 import android.graphics.Rect;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.recyclerview.widget.RecyclerView;
+
 import android.view.View;
 
 public class PaddingItemDecoration extends RecyclerView.ItemDecoration {
