@@ -9,8 +9,8 @@ import android.graphics.Bitmap;
 import android.net.Uri;
 import android.os.Environment;
 import android.provider.OpenableColumns;
-import android.support.v4.app.ShareCompat;
-import android.support.v4.content.FileProvider;
+import androidx.core.app.ShareCompat;
+import androidx.core.content.FileProvider;
 import android.util.Base64;
 import android.util.Log;
 import android.view.Gravity;
